// web/src/api/generated/model/userDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface UserDto {
  id: string;
  userName: string;
  displayName: string;
  roleId: string;
  roleNameEn: string;
  roleNameAr: string;
  isActive: boolean;
  mustChangePassword: boolean;
  /** @nullable */
  lastSignInAt: string | null;
}